Refactoring common ClojureScriptEngine code into a base class.
