+++ lib/clementine/clojurescript_engine/base.rb
@@ -1,0 +1,15 @@
+
+module Clementine
+  class ClojureScriptEngineBase
+    def default_opts
+      key = "output_dir"
+      value = ""
+      if defined?(Rails)
+        value = File.join(Rails.root, "app", "assets", "javascripts", "clementine")
+      else
+        value = Dir.pwd
+      end
+      {key => value}
+    end
+  end
+end
